+++ Runtime/Views/OptionView.cs
@@ -45,7 +45,7 @@ namespace Yarn.Unity
 
                 if (palette != null)
                 {
-                    text.text = LineView.PaletteMarkedUpText(line, palette);
+                    text.text = LineView.PaletteMarkedUpText(line, palette, false);
                 }
                 else
                 {